moved about, skills, projects, contact to components added components to index page added next-less plugin to handle styling
