+++ components/Contact.js
@@ -1,0 +1,13 @@
+const Contact = () => {
+	return (
+		<section>
+			<h1>Contact</h1>
+			<form>
+				Contact Form
+				<button>Send Message</button>
+			</form>
+		</section>
+	)
+};
+
+export default Contact;
