+++ components/Nav.js
@@ -7,16 +7,16 @@ const Nav = () => {
 				<h1>LOGO</h1>
 			</Link>
 			<nav>
-				<Link href="/about">
+				<Link href="#about">
 					<a>About Me</a>
 				</Link>
-				<Link href="/skills">
+				<Link href="#skills">
 					<a>Skills</a>
 				</Link>
-				<Link href="/projects">
+				<Link href="#projects">
 					<a>Projects</a>
 				</Link>
-				<Link href="/Contact">
+				<Link href="#contact">
 					<a>Contact</a>
 				</Link>
 			</nav>
