+++ components/Projects.js
@@ -1,0 +1,12 @@
+const Projects = () => {
+	return (
+		<section>
+			<h1>Projects</h1>
+			<div>
+				Projects
+			</div>
+		</section>
+	)
+};
+
+export default Projects;
