+++ components/Skills.js
@@ -1,0 +1,12 @@
+const Skills = () => {
+	return (
+		<section id="skills">
+			<h1>Skills</h1>
+			<div>
+				images
+			</div>
+		</section>
+	)
+};
+
+export default Skills;
