+++ package.json
@@ -20,7 +20,9 @@
   },
   "homepage": "https://github.com/jarata/portfolio-next#readme",
   "dependencies": {
+    "@zeit/next-less": "^1.0.1",
     "isomorphic-unfetch": "^3.0.0",
+    "less": "^3.10.3",
     "next": "^9.2.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0"
