--- pages/about.js
@@ -1,16 +1,0 @@
-import Layout from "../components/Layout";
-
-const About = () => {
-	return (
-		<Layout>
-			<main>
-				<h1>About Me</h1>
-				<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed sed ex tortor. Vivamus id nunc magna. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Donec in leo eget dui aliquam congue. Fusce vehicula, risus ut iaculis feugiat, risus mi congue nisl, non vehicula ipsum ipsum at neque. Nunc vel volutpat turpis. Phasellus in augue urna. Ut at metus sit amet risus vehicula scelerisque. Cras id laoreet est.</p>
-				<h2>Some other info</h2>
-				<p>Sed sodales rutrum tempus. Fusce vel venenatis mi. Quisque aliquet dignissim lacinia. Donec vel quam rutrum, maximus velit ut, mattis nunc. Vestibulum turpis magna, suscipit eget eros eget, efficitur vulputate elit. Praesent et diam placerat, blandit ante sit amet, scelerisque justo. Mauris cursus, justo at malesuada interdum, sapien lacus commodo odio, eu commodo dolor leo sed leo. Pellentesque scelerisque velit vitae eleifend gravida. Morbi efficitur tristique magna et condimentum. Nam a odio velit. Aliquam ex ex, sodales vitae ligula nec, eleifend sodales enim. Suspendisse nisl ligula, hendrerit eget lorem at, tempor hendrerit diam. Vestibulum sed justo eu eros bibendum commodo eu a sem. Fusce non vehicula nisl. Nam finibus, nisl sed egestas lacinia, purus quam convallis mi, vitae rhoncus augue purus vitae risus. Aliquam erat volutpat.</p>
-			</main>
-		</Layout>
-	)
-};
-
-export default About;
