--- pages/contact.js
@@ -1,13 +1,0 @@
-import Layout from "../components/Layout";
-
-const Contact = () => {
-	return (
-		<Layout>
-			<section>
-				Contact
-			</section>
-		</Layout>
-	)
-};
-
-export default Contact;
