--- pages/projects.js
@@ -1,13 +1,0 @@
-import Layout from "../components/Layout";
-
-const Projects = () => {
-	return (
-		<Layout>
-			<section>
-				Projects
-			</section>
-		</Layout>
-	)
-};
-
-export default Projects;
