--- pages/skills.js
@@ -1,13 +1,0 @@
-import Layout from "../components/Layout";
-
-const Skills = () => {
-	return (
-		<Layout>
-			<section>
-				Skills
-			</section>
-		</Layout>
-	)
-};
-
-export default Skills;
